Added chest fill in every world
