+++ Varo - 1.7 - 1.15/src/de/cuuky/varo/game/threads/VaroStartThread.java
@@ -1,16 +1,10 @@
 package de.cuuky.varo.game.threads;
 
 import java.net.InetAddress;
-import java.util.ArrayList;
-import java.util.List;
 
 import org.bukkit.Bukkit;
 import org.bukkit.GameMode;
-import org.bukkit.Location;
-import org.bukkit.block.Block;
-import org.bukkit.block.Chest;
 import org.bukkit.entity.Player;
-import org.bukkit.inventory.ItemStack;
 
 import de.cuuky.varo.Main;
 import de.cuuky.varo.api.VaroAPI;
@@ -21,9 +15,9 @@ import de.cuuky.varo.entity.player.VaroPlayer;
 import de.cuuky.varo.game.VaroGame;
 import de.cuuky.varo.game.start.ProtectionTime;
 import de.cuuky.varo.game.state.GameState;
+import de.cuuky.varo.game.world.VaroWorld;
 import de.cuuky.varo.game.world.border.decrease.BorderDecreaseMinuteTimer;
 import de.cuuky.varo.logger.logger.EventLogger.LogType;
-import de.cuuky.varo.utils.JavaUtils;
 import de.cuuky.varo.version.VersionUtils;
 import de.cuuky.varo.version.types.Sounds;
 
@@ -36,53 +30,6 @@ public class VaroStartThread implements Runnable {
 		this.game = Main.getVaroGame();
 
 		loadVaraibles();
-	}
-
-	private void fillChests() {
-		if(!ConfigSetting.RANDOM_CHEST_FILL_RADIUS.isIntActivated())
-			return;
-
-		int radius = ConfigSetting.RANDOM_CHEST_FILL_RADIUS.getValueAsInt();
-		Location loc = Main.getVaroGame().getVaroWorldHandler().getMainWorld().getWorld().getSpawnLocation().clone().add(radius, radius, radius);
-		Location loc2 = Main.getVaroGame().getVaroWorldHandler().getMainWorld().getWorld().getSpawnLocation().clone().add(-radius, -radius, -radius);
-
-		int itemsPerChest = ConfigSetting.RANDOM_CHEST_MAX_ITEMS_PER_CHEST.getValueAsInt();
-		ArrayList<ItemStack> chestItems = Main.getDataManager().getListManager().getChestItems().getItems();
-		for(Block block : getBlocksBetweenPoints(loc, loc2)) {
-			if(!(block.getState() instanceof Chest))
-				continue;
-
-			Chest chest = (Chest) block.getState();
-			chest.getBlockInventory().clear();
-			for(int i = 0; i < itemsPerChest; i++) {
-				int random = JavaUtils.randomInt(0, chest.getBlockInventory().getSize() - 1);
-				while(chest.getBlockInventory().getContents().length != chest.getBlockInventory().getSize())
-					random = JavaUtils.randomInt(0, chest.getBlockInventory().getSize() - 1);
-
-				chest.getBlockInventory().setItem(random, chestItems.get(JavaUtils.randomInt(0, chestItems.size() - 1)));
-			}
-		}
-
-		Bukkit.broadcastMessage("§7Alle Kisten um den " + Main.getColorCode() + "Spawn §7wurden " + Main.getColorCode() + "aufgefuellt§7!");
-	}
-
-	private List<Block> getBlocksBetweenPoints(Location l1, Location l2) {
-		List<Block> blocks = new ArrayList<>();
-		int topBlockX = (Math.max(l1.getBlockX(), l2.getBlockX()));
-		int bottomBlockX = (Math.min(l1.getBlockX(), l2.getBlockX()));
-		int topBlockY = (Math.max(l1.getBlockY(), l2.getBlockY()));
-		int bottomBlockY = (Math.min(l1.getBlockY(), l2.getBlockY()));
-		int topBlockZ = (Math.max(l1.getBlockZ(), l2.getBlockZ()));
-		int bottomBlockZ = (Math.min(l1.getBlockZ(), l2.getBlockZ()));
-
-		for(int x = bottomBlockX; x <= topBlockX; x++) {
-			for(int y = bottomBlockY; y <= topBlockY; y++) {
-				for(int z = bottomBlockZ; z <= topBlockZ; z++) {
-					blocks.add(l1.getWorld().getBlockAt(x, y, z));
-				}
-			}
-		}
-		return blocks;
 	}
 
 	public void loadVaraibles() {
@@ -150,9 +97,10 @@ public class VaroStartThread implements Runnable {
 			this.startcountdown = ConfigSetting.STARTCOUNTDOWN.getValueAsInt();
 			this.game.setMinuteTimer(new BorderDecreaseMinuteTimer());
 
-			fillChests();
+			for(VaroWorld world : Main.getVaroGame().getVaroWorldHandler().getWorlds())
+				world.fillChests();
+			
 			Main.getVaroGame().getVaroWorldHandler().getMainWorld().getWorld().strikeLightningEffect(Main.getVaroGame().getVaroWorldHandler().getMainWorld().getWorld().getSpawnLocation());
-
 			Bukkit.broadcastMessage(ConfigMessages.GAME_VARO_START.getValue());
 			Main.getDataManager().getVaroLoggerManager().getEventLogger().println(LogType.ALERT, ConfigMessages.ALERT_GAME_STARTED.getValue());
 			Bukkit.getScheduler().cancelTask(game.getStartScheduler());
